django_demo/pyproject.toml
[project]
name = "sapl-django-demo"
version = "1.0.0"
description = "Django demo application for SAPL Python PEP"
requires-python = ">=3.12"
dependencies = [
    "django>=5.0",
    "sapl-django>=4.0.0a1",
    "pyjwt[crypto]>=2.8",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "uvicorn[standard]>=0.27",
    "structlog>=24.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
